% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_summary.R
\name{cohort_summary}
\alias{cohort_summary}
\title{cohort_summary}
\usage{
cohort_summary(cohort_object)
}
\arguments{
\item{cohort_object}{output object of the create_cohort function.}
}
\description{
Creates a summary of the cohort created.
}
