% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohort.R
\name{create_cohort}
\alias{create_cohort}
\title{create_cohort}
\usage{
create_cohort(
  cohort = c("NSCLC", "CRC"),
  institution = c("DFCI", "MSK", "VICC", "UHN"),
  stage_dx,
  ca_hist_adeno_squamous,
  index_ca_dx = "first",
  regimen_drugs,
  regimen_number = "first",
  return_summary = FALSE
)
}
\arguments{
\item{cohort}{GENIE BPC Project cancer}

\item{institution}{GENIE BPC participating institution}

\item{stage_dx}{Stage at diagnosis}

\item{ca_hist_adeno_squamous}{Cancer histology}

\item{index_ca_dx}{Order of index cancer (first index cancer or all index cancers)}

\item{regimen_drugs}{Names of drugs in cancer-directed regimen}

\item{regimen_number}{Order of cancer-directed regimens returned (first time a regimen was received or all times a regimen was received)}

\item{return_summary}{Specifies whether or not a summary table for the cohort is returned}
}
\value{

}
\description{
This function allows the user to create a relevant cohort for analysis based on cancer diagnosis information such as cancer cohort, treating institution, histology, stage at diagnosis, as well as cancer-directed regimen information including regimen name and regimen order.
This will return a dataset (`cohort_ca_dx`) containing all patients matching the criterion of interest. If drug regimen criteria are also specified, then a second dataset (`cohort_ca_drugs`) will also be returned.
}
