% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_genomic_surv.R
\name{create_genomic_surv}
\alias{create_genomic_surv}
\title{create_genomic_surv}
\usage{
create_genomic_surv(
  samples_object,
  gen_dat,
  surv_type = "pfs_m",
  ignore_pb_LT = ifelse(surv_type \%in\% "os", FALSE, TRUE),
  complete_cases = TRUE
)
}
\arguments{
\item{samples_object}{output object of the opt_samples function (also works with fetch_samples function).}

\item{gen_dat}{output of binmat() (from package gnomeR) created from the corresponding samples.}

\item{surv_type}{select a time to event type of interest. Options are "os", "pfs_i" or "pfs_m".
Default is "pfs_m".}

\item{ignore_pb_LT}{boolean specifying if cases with time1 > time2 should be treated as no delayed truncation.
Default is TRUE if event type is PFS, FALSE if event type is OS.}

\item{complete_cases}{boolean specifying if only complete cases should be kept. Default is TRUE.}
}
\description{
Create a time to event dataset with their corresponding features
}
