% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.appen_abc}
\alias{.appen_abc}
\title{Reformat appendix tables to alphabetized two column pages}
\usage{
.appen_abc(rows_per_page, data)
}
\arguments{
\item{rows_per_page}{number of rows per page desired}

\item{data}{an appendix data frame to alphabetize}
}
\value{
a list of data frames (one for every page) where appendix
        tables start on the left column, reach the end of the page,
        and pick up again on the right column in alphabetical order
}
\description{
Reformat appendix tables to alphabetized two column pages
}
\keyword{internal}
