% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.bind_genie_data}
\alias{.bind_genie_data}
\title{Bind together a list of synapse data objects}
\usage{
.bind_genie_data(data_list)
}
\arguments{
\item{data_list}{a list of synapse data objects}
}
\value{
a list of GENIE dataframes (pt_char, ca_dx, etc.) that contains
        data from more than one cohort
}
\description{
Bind together a list of synapse data objects
}
\keyword{internal}
