% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_synapse.R
\name{.check_download_path}
\alias{.check_download_path}
\title{Check download_path user passed and create folder if needed}
\usage{
.check_download_path(download_location, version_num)
}
\arguments{
\item{download_location}{a local path or NULL}

\item{version_num}{vector of cohort/version_number}
}
\value{
a vector of file paths. If download_location is NULL, will return
temporary file path
}
\description{
Check download_path user passed and create folder if needed
}
\examples{
.check_download_path(download_location = NULL, version_num = "CRC_v2.1")

}
\keyword{internal}
