% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_authentication.R
\name{.check_synapse_login}
\alias{.check_synapse_login}
\title{Check Synapse Login Status & Ability to Access Data}
\usage{
.check_synapse_login()
}
\value{
Returns message indicating user is logged into
Synapse and has permission to access the GENIE BPC data.
}
\description{
The `.check_synapse_login()` function assesses whether the
user is logged into Synapse and confirms whether the
user has permission to access the GENIE BPC data.
}
\examples{
.check_synapse_login()
}
\keyword{internal}
