% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_authentication.R
\name{.get_token_by_pat}
\alias{.get_token_by_pat}
\title{Retrieve a synapse token using personal access token}
\usage{
.get_token_by_pat(pat)
}
\arguments{
\item{pat}{'Synapse' Personal Access Token. If NULL, package will search package
environment for "pat".}
}
\value{
a 'Synapse' token
}
\description{
Retrieve a synapse token using personal access token
}
\examples{
\dontrun{
.get_token_by_pat(pat = "test")
}
}
\keyword{internal}
