% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_authentication.R
\name{.get_token_by_username}
\alias{.get_token_by_username}
\title{Retrieve a synapse token using username and password}
\usage{
.get_token_by_username(username, password)
}
\arguments{
\item{username}{'Synapse' username. If NULL, package will search package
environment for "username".}

\item{password}{'Synapse' password. If NULL, package will search package
environment for "password".}
}
\value{
a 'Synapse' token.
}
\description{
Retrieve a synapse token using username and password
}
\examples{
\dontrun{
.get_token_by_username(username = "test", password = "test")
}
}
\keyword{internal}
