% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_synapse.R
\name{.pull_by_synapse_ids}
\alias{.pull_by_synapse_ids}
\title{Function to retrieve data by synapse ID}
\usage{
.pull_by_synapse_ids(synapse_ids_df, token, download_location)
}
\arguments{
\item{synapse_ids_df}{a dataframe of synapse IDs}

\item{token}{a synapse token}

\item{download_location}{if `NULL` (default), data will be returned as a list of dataframes with
requested data as list items. Otherwise, specify a folder path to have data automatically downloaded there.}
}
\value{
downloaded synapse data as a list if `download_location`= `NULL, or to a local path
}
\description{
Function to retrieve data by synapse ID
}
\examples{
syn_df <- data.frame(
  synapse_id =
    c("syn26046793", "syn26046791", "syn26046792")
)

.pull_by_synapse_ids(
  synapse_ids_df = syn_df,
  token = .get_synapse_token(), download_location = NULL
)

.pull_by_synapse_ids(
  synapse_ids_df = syn_df,
  token = .get_synapse_token(), download_location = here::here()
)

}
\keyword{internal}
