% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_authentication.R
\name{.verify_pat_works}
\alias{.verify_pat_works}
\title{Verify a synapse Personal Access Token works}
\usage{
.verify_pat_works(pat)
}
\arguments{
\item{pat}{'Synapse' Personal Access Token. If NULL, package will search package
environment for "pat".}
}
\value{
a 'Synapse' token
}
\description{
Verify a synapse Personal Access Token works
}
\examples{
\dontrun{
.verify_pat_works(pat = "test")
}
}
\keyword{internal}
