% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_names_by_cohort}
\alias{drug_names_by_cohort}
\title{Drug Names by Cohort}
\format{
A table for cancer-directed drug names associated with each cancer cohort:
\describe{
  \item{cohort}{GENIE BPC Project cancer. Must be one of "NSCLC" (non-small cell lung cancer) or "CRC" (colorectal cancer). Future cohorts will include "BrCa" (breast cancer), "PANC" (pancreatic cancer), "Prostate" (prostate cancer).}
  \item{drug_name}{Name of generic/ingredient cancer-directed drug}
  \item{drug_name_full}{Name of generic/ingredient cancer-directed drug with associated synonyms in parentheses}
  \item{immunotherapy}{Indicator for whether cancer-directed drug is an immunotherapy}#'   ...
}
}
\usage{
drug_names_by_cohort
}
\description{
A dataset containing the cancer-directed drug names and their synonyms.
}
\keyword{datasets}
