% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_regimen_list}
\alias{drug_regimen_list}
\title{List of Drug Regimen Names by Cohort}
\format{
A table for cancer-directed drug names associated with each
data release:
\describe{
  \item{cohort}{GENIE BPC Project cancer. One of "NSCLC"
  (non-small cell lung cancer), "CRC" (colorectal cancer),
  "BrCa" (breast cancer), "PANC" (pancreatic cancer),
  "Prostate" (prostate cancer), "BLADDER" (bladder cancer),
  and "RENAL" (renal cancer).}
  \item{cohort_data_release}{GENIE BPC data release. Occasionally, drug names
  were updated across releases to include additional drug name synonyms.}
  \item{drug_name}{Name of generic/ingredient cancer-directed drug}
  \item{drug_name_full}{Name of generic/ingredient cancer-directed drug
  with associated synonyms in parentheses}
  ...
}
}
\usage{
drug_regimen_list
}
\description{
A dataset containing the cancer-directed drug names and their synonyms.
}
\keyword{datasets}
