% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_genomics.R
\name{fetch_genomics}
\alias{fetch_genomics}
\title{fetch_genomics}
\usage{
fetch_genomics(samples_object, maf, fusion, cna)
}
\arguments{
\item{samples_object}{output object of the opt_samples function (also works with fetch_samples function).}

\item{maf}{maf file to be used for mutations}

\item{fusion}{fusion file to be used for fusions}

\item{cna}{cna file to be used for cna}
}
\description{
Creates mutation, fusion and cna files corresponding to the samples of interest by intersecting gene panels
from the different institutions.
}
