% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_samples.R
\name{fetch_samples}
\alias{fetch_samples}
\title{Link patients with corresponding next-generation sequencing reports}
\usage{
fetch_samples(data_synapse, df_record_ids)
}
\arguments{
\item{data_synapse}{The item from the nested list returned from
`pull_data_synapse()`}

\item{df_record_ids}{NGS data frame from the `create_analytic_cohort()`
function. Must contain variables: cohort, record_id and ca_seq.}
}
\value{
The NGS reports for each patient, stored as the
  'cohort_ngs' data frame of the create_analytic_cohort object
}
\description{
This function links patients in a cohort (created by
create_analytic_cohort()) with their corresponding
next generation sequencing (NGS) reports, which can be used to link to the
corresponding GENIE genomic data.
}
\details{
Subset cancer panel test data to patients in the cohort of interest
}
\author{
Axel Martin, Hannah Fuchs, Michael Curry
}
