% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genie_panels}
\alias{genie_panels}
\title{Genomic Panels Included in GENIE BPC Data}
\format{
A data frame with 12 rows and 3 variables:
\describe{
  \item{Sequence.Assay.ID}{Next-generation sequencing targeted panel assay identifier}
  \item{Panel}{Panel name}
  \item{Genes}{Number of genes included}
  ...
}
}
\usage{
genie_panels
}
\description{
A dataset containing the name, assay identifier, and number of genes
in each next-generation sequencing targeted panel included in GENIE BPC.
}
\keyword{datasets}
