% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsclc_test_data}
\alias{nsclc_test_data}
\title{Simulated fake GENIE BPC data for function examples and tests}
\format{
A list of data frames
\describe{
    \item{pt_char}{Patient characteristic data.frame}
    \item{ca_dx_index}{Index cancer diagnosis data.frame}
    \item{ca_dx_non_index}{Non-index cancer diagnosis data.frame}
    \item{ca_drugs}{Cancer directed-regimen data.frame}
    \item{prissmm_imaging}{PRISSMM Imaging report data.frame}
    \item{prissmm_pathology}{PRISSMM Pathology report data.frame}
    \item{prissmm_md}{PRISSMM medical oncologist report data.frame}
    \item{cpt}{Cancer Panel Test (CPT)/Next Generation Sequencing (NGS) data.frame}
    \item{mutations_extended}{Mutations data.frame}
    \item{fusions}{Fusions data.frame}
    \item{cpt}{Copy Number Alteration (CNA) data.frame}
}
}
\usage{
nsclc_test_data
}
\description{
A named list of simulated NSCLC clinical and genomic data
}
\keyword{datasets}
