% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_samples.R
\name{opt_samples}
\alias{opt_samples}
\title{opt_samples}
\usage{
opt_samples(
  samples_object,
  sample_type = c("Primary", "Metastasis"),
  histology
)
}
\arguments{
\item{samples_object}{output object of the fectch_samples function.}

\item{sample_type}{character specifying which type of sample to prioritize, options are "Primary" and "Metastasis".
Default is either.}

\item{histology}{character vector specifying which sample histologies to keep. See "Oncotree.Code" column
of samples_object argument above to get options.}
}
\description{
Get the best samples for analysis following several user define criterions.
}
