% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimen_abbreviations.R
\docType{data}
\name{regimen_abbreviations}
\alias{regimen_abbreviations}
\title{List of Drug Regimen Abbreviations}
\format{
A table for cancer-directed drug regimens and their common abbreviations
\describe{
  \item{regimen_drugs}{List of all drugs in the regimen}
  \item{abbreviation}{Common name of drug regimen, e.g. FOLFOX}
  ...
}
}
\usage{
regimen_abbreviations
}
\description{
A dataset containing the cancer-directed drug regimens and their common names
}
\keyword{datasets}
