% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_summary.R
\name{sample_summary}
\alias{sample_summary}
\title{sample_summary}
\usage{
sample_summary(samples_object)
}
\arguments{
\item{samples_object}{output object of the fectch_samples function.}
}
\description{
Creates a summary of the samples retrieved.
}
