% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synapse_tables}
\alias{synapse_tables}
\title{'Synapse' table IDs}
\format{
A lookup table for 'Synapse' data table IDs:
\describe{
  \item{cohort}{GENIE BPC Project Cohort}
  \item{version}{Release version}
  \item{release_date}{Month and year of data release}
  \item{df}{Clinical, Genomic, or Metadata Dataset}
  \item{synapse_id}{'Synapse' table ID for each dataset}
  ...
}
}
\source{
\url{https://www.synapse.org/#!Synapse:syn21226493/wiki/599164}
}
\usage{
synapse_tables
}
\description{
A dataset containing the 'Synapse' table IDs for
each dataset in GENIE BPC.
}
\keyword{datasets}
