% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment-history.R
\name{treatment_history}
\alias{treatment_history}
\title{treatment_history}
\usage{
treatment_history(ids, ca_drugs, regimen_drugs, lines_keep = NULL)
}
\arguments{
\item{ids}{dataframe with first column `record_id` and second column `ca_seq`.}

\item{ca_drugs}{dataset `ca_drugs` from `pull_data_synapse()` function.}

\item{regimen_drugs}{Vector with names of drugs in cancer-directed regimen,
separated by a comma. For example, to specify a regimen consisting of
Carboplatin and Pemetrexed, specify regimen_drugs = "Carboplatin,
Pemetrexed". Acceptable values are found in the `drug_names_by_cohort`
dataset provided with this package.}

\item{lines_keep}{regimen number to be kept to create the summary.}
}
\value{
Returns data frame `treat_hist` and interactive plot `p_dist`.
}
\description{
This function allows the user to get the complete treatment course of a list of patients.
}
\examples{
# pull_data_synapse("NSCLC")
# record_ids <- ca_dx_index_NSCLC$record_id[ca_dx_index_NSCLC$stage_dx == "Stage IV"]
# ca_drugs <- ca_drugs_NSCLC
# lines_keep = 1:3
# treatment_history(record_ids = record_ids, ca_drugs = ca_drugs,lines_keep = lines_keep)
}
