% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_authentication.R
\name{check_genie_access}
\alias{check_genie_access}
\title{Check Access to GENIE Data}
\usage{
check_genie_access(username = NULL, password = NULL)
}
\arguments{
\item{username}{Synapse username. If NULL, package will search package
environment for "username". If not found, package will look in
environmental variables for `SYNAPSE_USERNAME`.}

\item{password}{Synapse password. If NULL, package will search package
environment for "password". If not found package will search environmental
variables for `SYNAPSE_PASSWORD`.}
}
\value{
a success message if you are able to access genie data or a error
}
\description{
Check Access to GENIE Data
}
\examples{
\dontrun{
# if credentials are saved:
check_genie_access()
}
}
