% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cohort_institution}
\alias{cohort_institution}
\title{List of Institutions by Cohort}
\format{
A lookup table for institutions participating in curation by cancer cohort
\describe{
  \item{cohort}{GENIE BPC Project cancer. One of "NSCLC"
  (non-small cell lung cancer), "CRC" (colorectal cancer),
  "BrCa" (breast cancer), "PANC" (pancreatic cancer),
  "Prostate" (prostate cancer), and "BLADDER" (bladder cancer).}
  \item{institution}{GENIE BPC participating institution.}
  ...
}
}
\usage{
cohort_institution
}
\description{
A dataset containing the institutions participating in curation of each cancer cohort
}
\keyword{datasets}
