% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analytic_cohort.R
\name{create_analytic_cohort}
\alias{create_analytic_cohort}
\title{create_analytic_cohort}
\usage{
create_analytic_cohort(
  cohort,
  data_synapse,
  index_ca_seq = 1,
  institution,
  stage_dx,
  histology,
  regimen_drugs,
  regimen_type = "Exact",
  regimen_order,
  regimen_order_type,
  return_summary = FALSE
)
}
\arguments{
\item{cohort}{GENIE BPC Project cancer. Must be one of "NSCLC"
(non-small cell lung cancer) or "CRC" (colorectal cancer). Future cohorts
will include "BrCa" (breast cancer), "PANC" (pancreatic cancer),
"Prostate" (prostate cancer).}

\item{data_synapse}{The list returned from pull_data_synapse().}

\item{index_ca_seq}{Index cancer sequence. Default is 1, indicating the
patient's first index cancer. The index cancer is also referred to as the
BPC Project cancer in the GENIE BPC Analytic Data Guide; this is the
cancer that met the eligibility criteria for the project and was selected
at random for PRISSMM phenomic data curation. Specifying multiple index
cancer sequences, e.g. index_ca_seq = c(1, 2) will return index cancers to
patients with 1 index cancer and will return the first AND second index
cancers to patients with multiple.}

\item{institution}{GENIE BPC participating institution. Must be one of
"DFCI", "MSK", "UHN", or "VICC" for NSCLC cohorts; must be one of "DFCI",
"MSK", "VICC" for CRC. Default selection is all institutions.}

\item{stage_dx}{Stage at diagnosis. Must be one of "Stage I", "Stage II",
"Stage III", "Stage I-III NOS", "Stage IV". Default selection is all stages.}

\item{histology}{Cancer histology. For all cancer cohorts except for BrCa
(breast cancer), this corresponds to the variable `ca_hist_adeno_squamous` and
must be one of
"Adenocarcinoma", "Squamous cell", "Sarcoma", "Small cell carcinoma", "Carcinoma",
"Other histologies/mixed tumor". For BrCa, this corresponds to the variable
`ca_hist_brca` and must be one of
"Invasive lobular carcinoma", "Invasive ductal carcinoma", "Other histology".
Default selection is all histologies.}

\item{regimen_drugs}{Vector with names of drugs in cancer-directed regimen,
separated by a comma. For example, to specify a regimen consisting of
Carboplatin and Pemetrexed, specify regimen_drugs = "Carboplatin,
Pemetrexed". Acceptable values are found in the `drug_regimen_list`
dataset provided with this package.}

\item{regimen_type}{Indicates whether the regimen(s) specified in
`regimen_drugs` indicates the exact regimen to return, or if regimens
containing the drugs listed in `regimen_drugs` should be returned. Must be
one of "Exact" or "Containing". The default is "Exact".}

\item{regimen_order}{Order of cancer-directed regimen. If multiple drugs
are specified, `regimen_order` indicates the regimen order for all drugs;
different values of `regimen_order` cannot be specified for different drug
regimens. If multiple values are specified, e.g. c(1, 2), then drug regimens
that met either order criteria are returned.}

\item{regimen_order_type}{Specifies whether the `regimen_order` parameter
refers to the order of receipt of the drug regimen within the cancer
diagnosis (across all other drug regimens; "within cancer") or the order of
receipt of the drug regimen within the times that that drug regimen was
administered (e.g. the first time carboplatin pemetrexed was received, out
of all times that the patient received carboplatin pemetrexed; "within
regimen"). Acceptable values are "within cancer" and "within regimen".}

\item{return_summary}{Specifies whether a summary table for the cohort is
returned. Default is FALSE. The `gtsummary` package is required to return a
summary table.}
}
\value{
Returns data frames `cohort_ca_dx` and `cohort_ca_drugs`
}
\description{
This function allows the user to create a cohort from the GENIE BPC data
based on cancer diagnosis information such as cancer cohort, treating
institution, histology, and stage at diagnosis, as well as cancer-directed
regimen information including regimen name and regimen order.
This function returns two datasets:
(1) `cohort_ca_dx` will contain cancer diagnosis information for patients
matching the specified criteria. This dataset is structured as one record per
 patient per associated cancer diagnosis.
(2) `cohort_ca_drugs` will return the drug-regimen information for patients
matching the specified criteria. This dataset is structured as one record per
 patient, per regimen and associated cancer diagnosis. For example, a drug
 regimen associated with treating two cancer diagnoses will have two records
 in this dataset.
The function inputs `cohort`, `institution`, `stage_dx`,
and `regimen_drugs` correspond to the variable
names in the GENIE BPC Analytic Data Guide, available on
\href{https://www.synapse.org/#!Synapse:syn21241322}{Synapse}.
}
\examples{
# Example 1 ----------------------------------
# Create a cohort of all patients with stage IV NSCLC adenocarcinoma and
# also return all of their corresponding cancer-directed drugs
# pull_data_synapse("NSCLC")
# create_analytic_cohort(cohort = "NSCLC",
#   stage_dx = c("Stage IV"),
#   histology = "Adenocarcinoma")

# Example 2 ----------------------------------
# Create a cohort of all NSCLC patients who received Cisplatin,
# Pemetrexed Disodium or Cisplatin, Etoposide as their first drug regimen
# for their first index NSCLC
# pull_data_synapse("NSCLC")
# create_analytic_cohort(cohort = "NSCLC",
#   regimen_drugs = c("Cisplatin, Pemetrexed Disodium", "Cisplatin,
#   Etoposide"),
#   regimen_order = 1,
#   regimen_order_type = "within cancer")

# Example 3 ----------------------------------
# Create a cohort of all NSCLC patients who received Cisplatin, Pemetrexed
# Disodium
# at any time throughout the course of treatment for their cancer diagnosis,
# but in the event that the patient received the drug multiple times,
# only select the first time.
# pull_data_synapse("NSCLC")
# create_analytic_cohort(cohort = "NSCLC",
#   regimen_drugs = c("Cisplatin, Pemetrexed Disodium"),
#   regimen_order = 1,
#   regimen_order_type = "within regimen")
}
\author{
Jessica Lavery
}
