% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohort.R
\name{create_cohort}
\alias{create_cohort}
\title{create_cohort}
\usage{
create_cohort(
  cohort,
  output_suffix,
  index_ca_seq = 1,
  institution = c("DFCI", "MSK", "VICC", "UHN"),
  stage_dx,
  ca_hist_adeno_squamous,
  mets_dx_stage_iv,
  regimen_drugs,
  regimen_order,
  regimen_order_type,
  return_summary = FALSE
)
}
\arguments{
\item{cohort}{GENIE BPC Project cancer. Must be one of "NSCLC" or "CRC".}

\item{output_suffix}{Suffix to append to returned data frames (cohort_ca_dx_suffix, cohort_ca_drugs_suffix)}

\item{index_ca_seq}{Index cancer sequence. Default is 1, indicating the patient's first index cancer.}

\item{institution}{GENIE BPC participating institution. Must be one of "DFCI", "MSK", "UHN", or "VICC" for NSCLC cohorts; must be one of "DFCI", "MSK", "VICC" for CRC. Default is all institutions.}

\item{stage_dx}{Stage at diagnosis. Must be one of "Stage I", "Stage II", "Stage III", "Stage I-III NOS", "Stage IV". Default is all stages.}

\item{ca_hist_adeno_squamous}{Cancer histology. Must be one of "Adenocarcinoma", "Squamous cell", "Sarcoma", "Small cell carcinoma", "Other histologies/mixed tumor"}

\item{regimen_drugs}{Names of drugs in cancer-directed regimen, based on variable `regimen_drugs` in the ca_drugs dataset.}

\item{regimen_order}{Order of cancer-directed regimen. Default is 1, indicating the first time the regimen was received. If multiple drugs are specified, `regimen_order` indicates the regimen order for all drugs; different values of `regimen_order` cannot be specified for different drug regimens.}

\item{regimen_order_type}{Specifies whether the `regimen_order` parameter refers to the order of receipt of the drug regimen within the cancer diagnosis (across all other drug regimens) or the order of receipt of the drug regimen within the times that that drug regimen was administered (e.g. the first time carboplatin pemetrexed was received, out of all times that the patient received carboplatin pemetrexed)}

\item{return_summary}{Specifies whether a summary table for the cohort is returned. Default is FALSE.}
}
\value{

}
\description{
This function allows the user to create a cohort for analysis based on cancer diagnosis information such as cancer cohort, treating institution, histology, stage at diagnosis, as well as cancer-directed regimen information including regimen name and regimen order.
This will return two datasets: (1) cohort_ca_dx will contain all patients matching the cancer diagnosis criteria of interest; (2) cohort_ca_drugs will return the drug-regimen data to those patients.
}
