% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohort.R
\name{create_cohort}
\alias{create_cohort}
\title{create_cohort}
\usage{
create_cohort(
  cohort,
  institution = c("DFCI", "MSK", "VICC", "UHN"),
  stage_dx,
  ca_hist_adeno_squamous,
  index_ca_seq = 1,
  regimen_drugs,
  regimen_seq = 1,
  return_summary = FALSE
)
}
\arguments{
\item{cohort}{GENIE BPC Project cancer. Must be one of "NSCLC" or "CRC".}

\item{institution}{GENIE BPC participating institution. Must be one of "DFCI", "MSK", "UHN", or "VICC" for NSCLC cohorts; must be one of "DFCI", "MSK", "VICC" for CRC. Default is all institutions.}

\item{stage_dx}{Stage at diagnosis. Must be one of "Stage I", "Stage II", "Stage III", "Stage I-III NOS", "Stage IV". Default is all stages.}

\item{ca_hist_adeno_squamous}{Cancer histology. Must be one of "Adenocarcinoma", "Squamous cell", "Sarcoma", "Small cell carcinoma", "Other histologies/mixed tumor"}

\item{regimen_drugs}{Names of drugs in cancer-directed regimen, based on variable `regimen_drugs` in the ca_drugs dataset.}

\item{return_summary}{Specifies whether a summary table for the cohort is returned. Default is FALSE.}

\item{index_ca_dx}{Order of index cancer. Default is 1, indicating the first index cancer.}

\item{regimen_number}{Order of cancer-directed regimen. Default is 1, indicating the first time the regimen was received.}
}
\value{

}
\description{
This function allows the user to create a cohort for analysis based on cancer diagnosis information such as cancer cohort, treating institution, histology, stage at diagnosis, as well as cancer-directed regimen information including regimen name and regimen order.
This will return two datasets: (1) cohort_ca_dx will contain all patients matching the cancer diagnosis criteria of interest; (2) cohort_ca_drugs will return the drug-regimen data to those patients.
}
