% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_authentication.R
\name{.get_env}
\alias{.get_env}
\title{Get an object from the genieBPC package environment (`genieBPC_env`)}
\usage{
.get_env(thing_to_check)
}
\arguments{
\item{thing_to_check}{}
}
\value{
The object
}
\description{
Get an object from the genieBPC package environment (`genieBPC_env`)
}
\examples{
\dontrun{
.get_env("username")
}
}
\keyword{internal}
