% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_authentication.R
\name{.get_synapse_token}
\alias{.get_synapse_token}
\title{Retrieve a synapse token using PAT}
\usage{
.get_synapse_token(pat = NULL)
}
\arguments{
\item{pat}{'Synapse' Personal Access Token. If NULL, package will search package
environment for "pat".}
}
\value{
a 'Synapse' token. Will look for PAT and return that if successful HTTP call
}
\description{
Retrieve a synapse token using PAT
}
\examples{
\dontrun{
.get_synapse_token(pat = "test")
}
}
\keyword{internal}
