% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_authentication.R
\name{.get_synapse_token}
\alias{.get_synapse_token}
\title{Retrieve a synapse token using username and password}
\usage{
.get_synapse_token(username = NULL, password = NULL)
}
\arguments{
\item{username}{Synapse username. If NULL, package will search package
environment for "username". If not found, package will look in
environmental variables for `SYNAPSE_USERNAME`.}

\item{password}{Synapse password. If NULL, package will search package
environment for "password". If not found package will search environmental
variables for `SYNAPSE_PASSWORD`.}
}
\value{
a synapse token
}
\description{
Retrieve a synapse token using username and password
}
\examples{
\dontrun{
.get_synapse_token(username = "test", password = "test")
}
}
\keyword{internal}
