% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_authentication.R
\name{.is_connected_to_genie}
\alias{.is_connected_to_genie}
\title{Check Synapse Login Status & Ability to Access Data}
\usage{
.is_connected_to_genie(
  username = NULL,
  password = NULL,
  pat = NULL,
  check_consortium_access = FALSE
)
}
\value{
Returns message indicating user is logged into
'Synapse' and has permission to access the GENIE BPC data.
}
\description{
The `.is_connected_to_genie()` function assesses whether the
user is logged into 'Synapse' and confirms whether the
user has permission to access the GENIE BPC data. It returns TRUE or FALSE
}
\examples{
.is_connected_to_genie()
}
\keyword{internal}
