% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_synapse.R
\name{.pull_data_by_cohort}
\alias{.pull_data_by_cohort}
\title{Function to retrieve data by synapse ID}
\usage{
.pull_data_by_cohort(version_num_df, token, download_location)
}
\arguments{
\item{version_num_df}{a dataframe of synapse IDs}

\item{token}{a synapse token}

\item{download_location}{if `NULL` (default), data will be returned as a list of dataframes with
requested data as list items. Otherwise, specify a folder path to have data automatically downloaded there.}
}
\value{
downloaded synapse data as a list if `download_location`= `NULL, or to a local path
}
\description{
Function to retrieve data by synapse ID
}
\keyword{internal}
