% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_regimen_sunburst.R
\name{drug_regimen_sunburst}
\alias{drug_regimen_sunburst}
\title{drug_regimen_sunburst}
\usage{
drug_regimen_sunburst(data_synapse, data_cohort, max_n_regimens)
}
\arguments{
\item{max_n_regimens}{regimen number to be kept to create the summary.}

\item{ids}{dataframe with columns `record_id` and `ca_seq` from the downloaded data.}

\item{ca_drugs}{dataset `ca_drugs` from `pull_data_synapse()` function.}
}
\value{
Returns data frame `treatment_history` and interactive plot `sunburst_plot`.
}
\description{
This function allows the user to visualize the complete treatment course for selected diagnoses.
}
\examples{
# nsclc_data <- pull_data_synapse("NSCLC", version = "1.1")
# record_ids <- nsclc_data$ca_dx_index_NSCLC
# ca_drugs <- nsclc_data$ca_drugs_NSCLC
# test1 <- drug_regimen_sunburst(data_synapse = nsclc_data, data_cohort = ca_drugs,
max_n_regimens = 3)
}
