% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_regimen_sunburst.R
\name{drug_regimen_sunburst}
\alias{drug_regimen_sunburst}
\title{drug_regimen_sunburst}
\usage{
drug_regimen_sunburst(data_synapse, data_cohort, lines_keep = NULL)
}
\arguments{
\item{lines_keep}{regimen number to be kept to create the summary.}

\item{ids}{dataframe with columns `record_id` and `ca_seq` from the downloaded data.}

\item{ca_drugs}{dataset `ca_drugs` from `pull_data_synapse()` function.}
}
\value{
Returns data frame `treatment_history` and interactive plot `sunburst_plot`.
}
\description{
This function allows the user to visualize the complete treatment course for selected diagnoses.
}
\examples{
# nsclc_data <- pull_data_synapse("NSCLC", version = "1.1")
# record_ids <- nsclc_data$ca_dx_index_NSCLC
# ca_drugs <- nsclc_data$ca_drugs_NSCLC
# regimen_drugs <- unique(ca_drugs$regimen_drugs)
# lines_keep = 1:3
# test1 <- drug_regimen_sunburst(ids = record_ids, ca_drugs = ca_drugs,
#                            regimen_drugs = regimen_drugs, lines_keep = lines_keep)
}
