% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_regimen_sunburst.R
\name{drug_regimen_sunburst}
\alias{drug_regimen_sunburst}
\title{drug_regimen_sunburst}
\usage{
drug_regimen_sunburst(data_synapse, data_cohort, max_n_regimens = NULL)
}
\arguments{
\item{data_synapse}{The list returned from the `pull_data_synapse()`
function call}

\item{data_cohort}{The list returned from the `create_analytic_cohort()`
function call}

\item{max_n_regimens}{The number of regimens displayed in the sunburst plot}
}
\value{
Returns data frame `treatment_history` and interactive
plot `sunburst_plot`
}
\description{
This function allows the user to visualize the complete treatment
course for selected diagnoses.
}
\examples{
if(genieBPC:::.check_synapse_login() == TRUE){
nsclc_2_0 <- pull_data_synapse("NSCLC", version = "2.0-public")
nsclc_cohort <- create_analytic_cohort(data_synapse = nsclc_2_0$NSCLC_v2.0,
stage = "Stage IV")
ex1 <- drug_regimen_sunburst(data_synapse = nsclc_2_0$NSCLC_v2.0,
data_cohort = nsclc_cohort,
max_n_regimens = 3)
}
}
