% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_samples.R
\name{fetch_samples}
\alias{fetch_samples}
\title{fetch_samples}
\usage{
fetch_samples(clin_dat, cohort_object)
}
\arguments{
\item{clin_dat}{dataframe containing the sample information to be matched with the patients from the cohort of interest.
eg: data_clinical_sample.txt.}

\item{cohort_object}{output object of the create_cohort function.}
}
\description{
Add samples for each patients of the cohort of interest
}
