% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_samples.R
\name{fetch_samples}
\alias{fetch_samples}
\title{fetch_samples}
\usage{
fetch_samples(cohort, data_synapse, df_record_ids)
}
\arguments{
\item{cohort}{GENIE BPC Project cancer. Must be one of "NSCLC", "CRC",
or "BrCa"}

\item{data_synapse}{The item from the nested list returned from
`pull_data_synapse()`}

\item{df_record_ids}{NGS data frame from the `create_analytic_cohort()` function.}
}
\value{
returns the 'cohort_ngs' object of the create_analytic_cohort
with the genomic samples taken from each patients.
}
\description{
This function links patients in a cohort (created by
create_analytic_cohort()) with their corresponding
genomic samples available in GENIE.
}
\details{
Subset cancer panel test data to patients in the cohort of interest
}
\author{
Axel Martin
}
