% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_samples.R
\name{fetch_samples}
\alias{fetch_samples}
\title{fetch_samples}
\usage{
fetch_samples(cohort, cohort_object)
}
\arguments{
\item{cohort}{GENIE BPC Project cancer. Must be one of "NSCLC" or "CRC".}

\item{cohort_object}{output object of the create_cohort function.}
}
\value{
returns the cohort object list inputted with an additional dataset named "samples_data".
}
\description{
Add samples for each patients of the cohort of interest
}
\author{
Axel Martin
}
