% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_samples.R
\name{fetch_samples}
\alias{fetch_samples}
\title{fetch_samples}
\usage{
fetch_samples(cohort, data_synapse, df_record_ids)
}
\arguments{
\item{cohort}{GENIE BPC Project cancer. Must be one of "NSCLC", "CRC",
or "BrCa"}

\item{data_synapse}{the list object outputted by the
`pull_data_synapse()` function.}

\item{df_record_ids}{output object of the create_cohort function.}
}
\value{
returns the 'cohort_ngs' object of the create_analytic_cohort
with the genomic samples taken from each patients.
}
\description{
This function links patients in a cohort (created by
create_analytic_cohort()) with their corresponding
genomic samples available in GENIE.
}
\details{
Subset cancer panel test data to patients in the cohort of interest
}
\examples{
#' # Example 1 ----------------------------------
# Create a cohort of all patients with stage
# IV NSCLC of histology adenocarcinoma
# out <- create_analytic_cohort(cohort = "NSCLC",
# data_synapse = data_lung, stage_dx = c("Stage IV"),
# histology = "Adenocarcinoma")
# samples_data <- fetch_samples(cohort = "NSCLC",
# data_synapse = data_lung,df_record_ids = out$cohort_ngs)
# Example 2 ----------------------------------
# Create a cohort of all NSCLC patients who
# received Cisplatin, Pemetrexed Disodium or Cisplatin,
# Etoposide as their first drug regimen
# out <- create_analytic_cohort(cohort = "NSCLC",
# data_synapse = data_lung,regimen_drugs =
# c("Cisplatin, Pemetrexed Disodium",
# "Cisplatin, Etoposide"), regimen_order = 1,
# regimen_order_type = "within regimen")
# samples_data <- fetch_samples(cohort = "NSCLC",
# data_synapse = data_lung,df_record_ids = out$cohort_ngs)
}
\author{
Axel Martin
}
