% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genieBPC-package.R
\docType{package}
\name{genieBPC-package}
\alias{genieBPC}
\alias{genieBPC-package}
\title{genieBPC: Project GENIE BioPharma Collaborative Data Processing Pipeline}
\description{
The American Association Research (AACR) Project Genomics Evidence Neoplasia Information Exchange (GENIE) BioPharma Collaborative represents a multi-year, multi-institution effort to build a pan-cancer repository of linked clinico-genomic data. The genomic and clinical data are provided in multiple releases (separate releases for each cancer cohort with updates following data corrections), which are stored on the data sharing platform 'Synapse' \url{https://www.synapse.org/}. The 'genieBPC' package provides a seamless way to obtain the data corresponding to each release from 'Synapse' and to prepare datasets for analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://genie-bpc.github.io/genieBPC/}
  \item Report bugs at \url{https://github.com/GENIE-BPC/genieBPC/issues}
}

}
\author{
\strong{Maintainer}: Jessica A. Lavery \email{laveryj@mskcc.org} (\href{https://orcid.org/0000-0002-2746-5647}{ORCID})

Authors:
\itemize{
  \item Michael A. Curry (\href{https://orcid.org/0000-0002-0261-4044}{ORCID})
  \item Samantha Brown (\href{https://orcid.org/0000-0001-5352-974X}{ORCID})
  \item Karissa Whiting (\href{https://orcid.org/0000-0002-4683-1868}{ORCID})
  \item Axel Martin
}

Other contributors:
\itemize{
  \item Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID}) [contributor]
}

}
\keyword{internal}
