% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenieBPC-package.R
\docType{package}
\name{GenieBPC-package}
\alias{GenieBPC}
\alias{GenieBPC-package}
\title{GenieBPC: Helps format Genie data}
\description{
The {GenieBPC} package provides a seamless way to obtain the
    data corresponding to each release from Synapse and to prepare datasets
    for analysis.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/AxelitoMartin/GenieBPC/issues}
}

}
\author{
\strong{Maintainer}: Jessica Lavery \email{laveryj@mskcc.org} (\href{https://orcid.org/0000-0002-2746-5647}{ORCID})

Authors:
\itemize{
  \item Michael Curry (\href{https://orcid.org/0000-0002-0261-4044}{ORCID})
  \item Samantha Brown
  \item Axel Martin
}

}
\keyword{internal}
