% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsclc_test_data}
\alias{nsclc_test_data}
\title{Simulated fake synapse data for function examples and tests}
\format{
A list of cohort data frames
\describe{
    \item{pt_char_NSCLC}{Patient characteristic data.frame}
    \item{ca_dx_index_NSCLC}{Cancer Dx data.frame}
    \item{ca_drugs_NSCLC}{Drug data.frame}
    \item{cpt_NSCLC}{CPT data.frame}
}
}
\usage{
nsclc_test_data
}
\description{
A named list of simulated NSCLC data
}
\keyword{datasets}
