% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_synapse.R
\name{pull_data_synapse}
\alias{pull_data_synapse}
\title{pull_data_synapse}
\usage{
pull_data_synapse(cohort, version)
}
\arguments{
\item{cohort}{Specify the cohort of data to pull (NSCLC or CRC)}

\item{version}{Specify which version of the data to pull. By default, the most recent version is pulled.}
}
\description{
Pull data from synapse
}
