% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_synapse.R
\name{pull_data_synapse}
\alias{pull_data_synapse}
\title{Obtain Clinical Data Files for GENIE BPC Project}
\usage{
pull_data_synapse(cohort, version)
}
\arguments{
\item{cohort}{Vector or list specifying the cohort(s) of interest.
Must be one of "NSCLC" (Non-Small Cell Lung Cancer) or
 "CRC" (Colorectal Cancer).}

\item{version}{Vector or list specifying the version of the data.
By default, the most recent version is pulled.
Currently, only version 1.1 is available.
When entering multiple cohorts, the order of the
version numbers corresponds to the order that the cohorts
are specified; the cohort and version number must be in
the same order in order to pull the correct data. See examples below.}
}
\value{
Returns clinical and genomic data corresponding to
the specified cohort(s). Data frames have the suffix
indicating the cohort appended to their name,
e.g. pt_char_NSCLC for the pt_char dataset of
the NSCLC cohort.
}
\description{
The `pull_data_synapse` function accesses the specified
 version of the clinical and genomic GENIE BPC data from
 \href{https://www.synapse.org/#!Synapse:syn21226493/wiki/599164}{Synapse}
 and reads it into the R environment.
  Documentation corresponding to the clinical data files
  can also be found on Synapse in the Analytic Data Guide:
\itemize{
  \item \href{https://www.synapse.org/#!Synapse:syn23002641}{NSCLC v1.1-Consortium Analytic Data Guide}
  \item \href{https://www.synapse.org/#!Synapse:syn26008058}{NSCLC v2.1-Consortium Analytic Data Guide}
  \item \href{https://www.synapse.org/#!Synapse:syn23764204}{CRC v1.1-Consortium Analytic Data Guide}
  \item \href{https://www.synapse.org/#!Synapse:syn26077308}{CRC v1.2-Consortium Analytic Data Guide}
  \item \href{https://www.synapse.org/#!Synapse:syn26077313}{BrCa v1.1-Consortium Analytic Data Guide}
}
Users must log in to Synapse to access the data successfully.
To log into Synapse during each session, call:
`synLogin(email = "your_email", password = "your_password")`
To store authentication information in your operating system, call:
`synLogin(email="your_email",password="your_password",rememberMe=TRUE)`
Upon calling the `rememberMe = TRUE` argument, the user can call:
`synLogin()` in future uses without specifying login credentials.
}
\examples{
if(genieBPC:::.check_synapse_login() == TRUE){
# Example 1 ----------------------------------
# Pull non-small cell lung cancer data

 pull_data_synapse(cohort = "NSCLC", version = "2.1-consortium")

# Example 2 ----------------------------------
# Pull the most recent non-small cell lung cancer
# data and the most recent colorectal cancer data
 pull_data_synapse(cohort = c("NSCLC", "CRC"),
 version = c("2.1-consortium", "1.2-consortium"))

# Example 3 ----------------------------------
# Pull version 2.1 for non-small cell lung cancer
 #and version 1.1 for colorectal cancer data
 pull_data_synapse(
  cohort = c("NSCLC", "CRC"),
  version = c("2.1-consortium", "1.1-consortium")
 )
}
}
\author{
Michael Curry
}
