% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data_synapse.R
\name{pull_data_synapse}
\alias{pull_data_synapse}
\title{Obtain clinical & genomic data files for GENIE BPC Project}
\usage{
pull_data_synapse(
  cohort = NULL,
  version = NULL,
  download_location = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{cohort}{Vector or list specifying the cohort(s) of interest.
Must be one of "NSCLC" (Non-Small Cell Lung Cancer),
 "CRC" (Colorectal Cancer), or "BrCa" (Breast Cancer).}

\item{version}{Vector specifying the version of the data. Must be one of the following:
"v1.1-consortium", "v1.2-consortium", "v2.1-consortium", "v2.0-public". When entering multiple cohorts, the order of the
version numbers corresponds to the order that the cohorts
are specified; the cohort and version number must be in
the same order in order to pull the correct data. See examples below.}

\item{download_location}{if `NULL` (default), data will be returned as a list of dataframes with
requested data as list items. Otherwise, specify a folder path to have data automatically downloaded there.
When a path is specified, data are not read into the R environment.}

\item{username}{Synapse username}

\item{password}{Synapse password}
}
\value{
Returns a nested list of clinical and genomic data corresponding to
the specified cohort(s).
}
\description{
The `pull_data_synapse` function accesses the specified
 version of the clinical and genomic GENIE BPC data from
 \href{https://www.synapse.org/#!Synapse:syn21226493/wiki/599164}{Synapse}
 and reads it into the R environment.
  Documentation corresponding to the clinical data files
  can also be found on Synapse in the Analytic Data Guide:
\itemize{
  \item \href{https://www.synapse.org/#!Synapse:syn23002641}{NSCLC v1.1-Consortium Analytic Data Guide}
  \item \href{https://www.synapse.org/#!Synapse:syn26008058}{NSCLC v2.1-Consortium Analytic Data Guide}
  \item \href{https://www.synapse.org/#!Synapse:syn30557304}{NSCLC v2.0-Public Analytic Data Guide}
  \item \href{https://www.synapse.org/#!Synapse:syn23764204}{CRC v1.1-Consortium Analytic Data Guide}
  \item \href{https://www.synapse.org/#!Synapse:syn26077308}{CRC v1.2-Consortium Analytic Data Guide}
  \item \href{https://www.synapse.org/#!Synapse:syn26077313}{BrCa v1.1-Consortium Analytic Data Guide}
}
Users must log in to Synapse to access the data successfully.
To set your Synapse credentials during each session, call:
`set_synapse_credentials(username = "your_username", password = "your_password")`
To store authentication information in your environmental variables, add the
following to your .Renviron file (tip: you can use usethis::edit_r_environ() to easily open/edit this file):
`SYNAPSE_USERNAME = <your-username>`
`SYNAPSE_PASSWORD = <your-password>`
Alternatively, you can pass your username and password to each individual data pull function if preferred,
although it is recommended that you manage your passwords outside of your scripts for security purposes.
}
\details{
See the \href{https://genie-bpc.github.io/genieBPC/articles/pull_data_synapse_vignette.html}{pull_data_synapse vignette}
for further documentation and examples.
}
\examples{
\donttest{
if(genieBPC::.is_connected_to_genie()) {
# Example 1 ----------------------------------
# Pull non-small cell lung cancer data
set_synapse_credentials()
results <- pull_data_synapse(cohort = "NSCLC", version = "v2.1-consortium")
names(results$NSCLC_v2.1)

# Example 2 ----------------------------------
# Pull the most recent non-small cell lung cancer
# data and the most recent colorectal cancer data

 results2 <- pull_data_synapse(
   cohort = c("NSCLC", "CRC"),
   version = c("v2.1-consortium", "v1.2-consortium")
 )
names(results2)


# Example 3 ----------------------------------
# Pull version 2.1 for non-small cell lung cancer
# and version 1.1 for colorectal cancer data

 results3 <- pull_data_synapse(
   cohort = c("CRC", "BrCa"),
   version = c("v1.2-consortium", "v1.1-consortium")
 )
  str(results3)
  }
}

}
\author{
Karissa Whiting, Michael Curry
}
