% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_authentication.R
\name{set_synapse_credentials}
\alias{set_synapse_credentials}
\title{Connect to Synapse API}
\usage{
set_synapse_credentials(username = NULL, password = NULL)
}
\arguments{
\item{username}{Synapse username. If NULL, package will search environmental
variables for `SYNAPSE_USERNAME`.}

\item{password}{Synapse password. If NULL, package will search environmental
variables for `SYNAPSE_PASSWORD`.}
}
\description{
This function sets Synapse credentials for the user's current session
}
\examples{
\dontrun{
set_synapse_credentials(
  username = "your-username",
  password = "your-password"
)
}

}
