% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synapse_tables}
\alias{synapse_tables}
\title{Synapse table IDs}
\format{
A lookup table for Synapse clinical data table IDs:
\describe{
  \item{cohort}{GENIE BPC Project Cohort}
  \item{df}{Clinical dataset}
  \item{version}{Release version}
  \item{synapse_id}{Synapse table ID for each dataset}
  \item{release_date}{Month and year of data release}
  ...
}
}
\source{
\url{https://www.synapse.org/#!Synapse:syn21226493/wiki/599164}
}
\usage{
synapse_tables
}
\description{
A dataset containing the Synapse table IDs for
each clinical dataset in GENIE BPC.
}
\keyword{datasets}
